DJANGO/Blog/pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
